requirements.txt
google-auth 
google-api-python-client
oauthlib<3.2.2
requests
google-auth-httplib2 
google-auth-oauthlib